+++ CHANGELOG.md
@@ -1,17 +1,30 @@
 # Changelog
 
-## WIP
+## 3.4.3 - 2019-09-26
 
 ### Added
-- Added Craft 3.3 requirement
+- Added template hook `cp.sproutForms.forms.edit.content` ([#339pullrequest])
+- Added template hook `cp.sproutForms.forms.edit.details` ([#339pullrequest])
+- Added template hook `cp.sproutForms.entries.edit.content` ([#339pullrequest])
+- Added template hook `cp.sproutForms.entries.edit.details` ([#339pullrequest])
+
+### Changed
+- Updated craftcms/cms requirement to v3.3.0
 
 ### Fixed
-- Fixed bug where File Upload field doesn't update Asset Source option
+- Fixed deprecation error in default Email Template when using Relations fields. ([#90])
+- Fixed bug introduced in Craft v3.3.0 where File Upload field doesn't update Asset Source option ([#343])
+- Fixed status filter behavior ([#339pullrequest])
+- Fixed bug where it was not possible to create new Entry Statuses using Postgres
+
+[#90]: https://github.com/barrelstrength/craft-sprout-forms/issues/90
+[#343]: https://github.com/barrelstrength/craft-sprout-forms/issues/343
+[#339pullrequest]: https://github.com/barrelstrength/craft-sprout-forms/pull/339
 
 ## 3.4.2 - 2019-09-04
 
 ### Changed
-- Updated barrelstrength/sprout-base-reports requirement v1.3.8
+- Updated barrelstrength/sprout-base-reports requirement to v1.3.8
 
 ### Fixed
 - Fixed bug where field classes did not display in Basic Form Templates ([#335])